feat(run) : added CommandLineRunner instruction to verify bd functionnalities

+++ src/main/java/com/example/filrougefo/entity/OrderLine.java
@@ -17,7 +17,7 @@ public class OrderLine {
     @ManyToOne
     @JoinColumn(name = "idProduct")
     private Product product;
-    private int quantity;
+    private BigDecimal quantity;
     private BigDecimal discount;
 
 }
